+++ jotit/Program.cs
@@ -1,11 +1,12 @@
-﻿using JotIt.Data;
+﻿using JotIt;
+using JotIt.Data;
 using JotIt.Models;
 
 var repo = new ItemRepository("jotit.db");
 
 if (args.Length > 0)
 {
-    HandleCli(args);
+    new CliHandler(repo).Handle(args);
     return;
 }
 
@@ -88,30 +89,7 @@ string? GetCategory()
     return category;
 }
 
-string? GetDueDate(string? initial = null)
-{
-    string? input;
 
-    if (initial is not null)
-    {
-        input = initial; // CLI path: validate the provided value directly
-    }
-    else
-    {
-        // Interactive path: prompt the user
-        Console.Write($"Enter due date (yyyy-MM-dd) [{DateTime.Today:yyyy-MM-dd}]: ");
-        input = Console.ReadLine();
-        if (string.IsNullOrWhiteSpace(input))
-            return DateTime.Today.ToString("yyyy-MM-dd");
-    }
-
-    if (!DateTime.TryParseExact(input, "yyyy-MM-dd", null, System.Globalization.DateTimeStyles.None, out _))
-    {
-        Console.WriteLine("Invalid date format. Expected yyyy-MM-dd.");
-        return null;
-    }
-    return input;
-}
 
 void AddItem() {
     Console.WriteLine();
@@ -139,7 +117,7 @@ void AddItem() {
             string? body = GetBody();
             if (body == null) return;
             string? category = GetCategory();
-            string? dueDate = GetDueDate();
+            string? dueDate = Prompts.GetDueDate();
             if (dueDate == null) {
                 var note = NoteItem.Create(body, category);
                 if (note != null) { 
@@ -201,7 +179,7 @@ void ConvertNoteToTask(NoteItem note)
     Console.WriteLine();
     Console.WriteLine(note.ToString());
 
-    string? dueDate = GetDueDate();
+    string? dueDate = Prompts.GetDueDate();
     if (dueDate == null) return;
     repo.UpdateDueDate(note.Id, dueDate);
     Console.WriteLine("Note converted to task.");
@@ -259,169 +237,4 @@ void DeleteItem()
         Console.WriteLine("No item found with that ID.");
 }
 
-void HandleCli(string[] args)
-{
-    string verb = args[0].ToLower();
 
-    if (verb == "--help" || verb == "help")
-    {
-        PrintHelp();
-        return;
-    }
-
-    switch (verb)
-    {
-        case "add":
-            CliAdd(args[1..]);
-            break;
-        case "list":
-            CliList(args[1..]);
-            break;
-        case "delete":
-            CliDelete(args[1..]);
-            break;
-        case "change":
-            CliChange(args[1..]);
-            break;
-        default:
-            Console.WriteLine($"Unknown command: {verb}");
-            PrintHelp();
-            break;
-    }
-}
-
-void CliAdd(string[] args)
-{
-    if (args.Length == 0) { PrintHelp(); return; }
-
-    string type = args[0].ToLower();
-    var (body, flags) = ParseArgs(args[1..]);
-
-    if (string.IsNullOrWhiteSpace(body))
-    {
-        Console.WriteLine("Body cannot be empty.");
-        return;
-    }
-
-    string category = flags.GetValueOrDefault("--category", "");
-
-    switch (type)
-    {
-        case "note":
-            var note = NoteItem.Create(body, category);
-            if (note == null) return;
-            repo.Add(note);
-            Console.WriteLine("Note saved!");
-            break;
-        case "task":
-            string? dueDate = GetDueDate(flags.GetValueOrDefault("--due", DateTime.Today.ToString("yyyy-MM-dd")));
-            if (dueDate is null) return;
-            var task = TaskItem.Create(body, category, dueDate);
-            if (task == null) return;
-            repo.Add(task);
-            Console.WriteLine("Task saved!");
-            break;
-        default:
-            Console.WriteLine($"Unknown type: {args[0]}");
-            PrintHelp();
-            break;
-    }
-}
-
-void CliList(string[] args)
-{
-    string filter = args.Length > 0 ? args[0].ToLower() : "all";
-    switch (filter)
-    {
-        case "notes":
-            new NoteCollection(repo).Display();
-            break;
-        case "tasks":
-            new TaskCollection(repo).Display();
-            break;
-        default:
-            if (filter != "all") {
-                Console.WriteLine($"Unknown filter '{filter}', showing all.");
-            }
-            new NoteCollection(repo).Display();
-            new TaskCollection(repo).Display();
-            break;
-    }
-}
-
-void CliDelete(string[] args)
-{
-    if (args.Length == 0 || !long.TryParse(args[0], out long id))
-    {
-        Console.WriteLine("Usage: jotit delete <id>");
-        return;
-    }
-
-    if (repo.Delete(id))
-        Console.WriteLine("Item deleted.");
-    else
-        Console.WriteLine("No item found with that ID.");
-}
-
-void CliChange(string[] args)
-{
-    if (args.Length == 0 || !long.TryParse(args[0], out long id))
-    {
-        Console.WriteLine("Usage: jotit change <id> [--due yyyy-MM-dd]");
-        return;
-    }
-
-    var item = repo.GetItemById(id);
-    if (item == null) { Console.WriteLine("No item found with that ID."); return; }
-
-    var (_, flags) = ParseArgs(args[1..]);
-
-    if (item is NoteItem)
-    {
-        string? dueDate = GetDueDate(flags.GetValueOrDefault("--due", DateTime.Today.ToString("yyyy-MM-dd")));
-        if (dueDate is null) return;
-        repo.UpdateDueDate(id, dueDate);
-        Console.WriteLine("Note converted to task.");
-    }
-    else
-    {
-        repo.UpdateDueDate(id, null);
-        Console.WriteLine("Task converted to note.");
-    }
-}
-
-(string body, Dictionary<string, string> flags) ParseArgs(string[] args)
-{
-    var flags = new Dictionary<string, string>();
-    var bodyTokens = new List<string>();
-
-    for (int i = 0; i < args.Length; i++)
-    {
-        if (args[i].StartsWith("--") && i + 1 < args.Length)
-        {
-            flags[args[i]] = args[i + 1];
-            i++; // skip the value
-        }
-        else if (!args[i].StartsWith("--"))
-        {
-            bodyTokens.Add(args[i]);
-        }
-    }
-
-    return (string.Join(" ", bodyTokens), flags);
-}
-
-void PrintHelp()
-{
-    Console.WriteLine();
-    Console.WriteLine("Usage: jotit <command> [options]");
-    Console.WriteLine();
-    Console.WriteLine("Commands:");
-    Console.WriteLine("  add note <body> [--category <string>]");
-    Console.WriteLine("  add task <body> [--due yyyy-MM-dd] [--category <string>]");
-    Console.WriteLine("  list [notes|tasks]");
-    Console.WriteLine("  delete <id>");
-    Console.WriteLine("  change <id> [--due yyyy-MM-dd]");
-    Console.WriteLine();
-    Console.WriteLine("Run without arguments to launch the interactive menu.");
-}
